Refactor Query class into separate file
