+++ src/firecrypt.js
@@ -26,6 +26,7 @@ CryptoJS.enc.Base64UrlSafe = {
   var encryptString, decryptString;
 
   var utils = require('./utils');
+  var FireCryptQuery = require('./FireCryptQuery');
   var FireCryptSnapshot = require('./FireCryptSnapshot');
   var FireCryptOnDisconnect = require('./FireCryptOnDisconnect');
 
@@ -78,115 +79,6 @@ CryptoJS.enc.Base64UrlSafe = {
     if (checkValue) decryptString(checkValue);
     return encryptString(CryptoJS.enc.Base64UrlSafe.stringify(CryptoJS.lib.WordArray.random(10)));
   }
-
-  function Query(query, order, original) {
-    this._query = query;
-    this._order = order || {};
-    this._original = original || query;
-  }
-  Query.prototype.on = function(eventType, callback, cancelCallback, context) {
-    wrapQueryCallback(callback);
-    return this._original.on.call(
-      this._query, eventType, callback.firecryptCallback, cancelCallback, context);
-  };
-  Query.prototype.off = function(eventType, callback, context) {
-    if (callback && callback.firecryptCallback) callback = callback.firecryptCallback;
-    return this._original.off.call(this._query, eventType, callback, context);
-  };
-  Query.prototype.once = function(eventType, successCallback, failureCallback, context) {
-    wrapQueryCallback(successCallback);
-    return this._original.once.call(
-      this._query, eventType, successCallback && successCallback.firecryptCallback, failureCallback,
-      context
-    ).then(function(snap) {
-      return new FireCryptSnapshot(snap);
-    });
-  };
-  Query.prototype.orderByChild = function(key) {
-    return this._orderBy('orderByChild', 'child', key);
-  };
-  Query.prototype.orderByKey = function() {
-    return this._orderBy('orderByKey', 'key');
-  };
-  Query.prototype.orderByValue = function() {
-    return this._orderBy('orderByValue', 'value');
-  };
-  Query.prototype.orderByPriority = function() {
-    return this._orderBy('orderByPriority', 'priority');
-  };
-  Query.prototype.startAt = function(value, key) {
-    this._checkCanSort(key !== undefined);
-    return this._delegate('startAt', arguments);
-  };
-  Query.prototype.endAt = function(value, key) {
-    this._checkCanSort(key !== undefined);
-    return this._delegate('endAt', arguments);
-  };
-  Query.prototype.equalTo = function(value, key) {
-    if (this._order[this._order.by + 'Encrypted']) {
-      value = utils.encrypt(value, utils.getType(value), this._order[this._order.by + 'Encrypted']);
-    }
-    if (key !== undefined && this._order.keyEncrypted) {
-      key = utils.encrypt(key, 'string', this._order.keyEncrypted);
-    }
-    return new Query(this._original.equalTo.call(this._query, value, key), this._order);
-  };
-  Query.prototype.limitToFirst = function() {
-    return this._delegate('limitToFirst', arguments);
-  };
-  Query.prototype.limitToLast = function() {
-    return this._delegate('limitToLast', arguments);
-  };
-  Query.prototype.limit = function() {
-    return this._delegate('limit', arguments);
-  };
-  Query.prototype.ref = function() {
-    return utils.decryptRef(this._original.ref.call(this._query));
-  };
-  Query.prototype._delegate = function(methodName, args) {
-    return new Query(this._original[methodName].apply(this._query, args), this._order);
-  };
-  Query.prototype._checkCanSort = function(hasExtraKey) {
-    if (this._order.by === 'key' ?
-        this._order.keyEncrypted :
-        this._order.valueEncrypted || hasExtraKey && this._order.keyEncrypted) {
-      throw new Error('Encrypted items cannot be ordered');
-    }
-  };
-  Query.prototype._orderBy = function(methodName, by, childKey) {
-    var def = utils.specForPath(utils.refToPath(this.ref()));
-    var order = {by: by};
-    var encryptedChildKey;
-    if (def) {
-      var childPath = childKey && childKey.split('/');
-      for (var subKey in def) {
-        if (!def.hasOwnProperty(subKey)) continue;
-        var subDef = def[subKey];
-        if (subDef['.encrypt']) {
-          if (subDef['.encrypt'].key) order.keyEncrypted = subDef['.encrypt'].key;
-          if (subDef['.encrypt'].value) order.valueEncrypted = subDef['.encrypt'].value;
-        }
-        if (childKey) {
-          var childDef = utils.specForPath(childPath, subDef);
-          if (childDef && childDef['.encrypt'] && childDef['.encrypt'].value) {
-            order.childEncrypted = childDef['.encrypt'].value;
-          }
-          var encryptedChildKeyCandidate = utils.encryptPath(childPath, subDef).join('/');
-          if (encryptedChildKey && encryptedChildKeyCandidate !== encryptedChildKey) {
-            throw new Error(
-              'Incompatible encryption specifications for orderByChild("' + childKey + '")');
-          }
-          encryptedChildKey = encryptedChildKeyCandidate;
-        }
-      }
-    }
-    if (childKey) {
-      return new Query(
-        this._original[methodName].call(this._query, encryptedChildKey || childKey), order);
-    } else {
-      return new Query(this._original[methodName].call(this._query), order);
-    }
-  };
 
   function wrapFirebase() {
     if (firebaseWrapped) return;
@@ -279,17 +171,8 @@ CryptoJS.enc.Base64UrlSafe = {
   function interceptQuery(methodName) {
     originalQueryFbp[methodName] = fbp[methodName];
     fbp[methodName] = function() {
-      var query = new Query(utils.encryptRef(this), {}, originalQueryFbp);
+      var query = new FireCryptQuery(utils.encryptRef(this), {}, originalQueryFbp);
       return query[methodName].apply(query, arguments);
     };
-  }
-
-  function wrapQueryCallback(callback) {
-    if (!callback || callback.firecryptCallback) return;
-    var wrappedCallback = function(snap, previousChildKey) {
-      return callback.call(this, new FireCryptSnapshot(snap), previousChildKey);
-    };
-    wrappedCallback.firecryptCallback = wrappedCallback;
-    callback.firecryptCallback = wrappedCallback;
   }
 })();
